Cleanup: hspec top-level decls.
weird things have been observed to happen if there is more
than the top-level test case of the form `tests :: Spec = hspec specs.`
in particular, if it has the form `hspec specs1 >> hspec specs2`, or
`hspec $ specs1 >> specs2`, not all tests are always called when
running `cabal test`.

i don't have a good explanation or a way to reproduce this, but i
suggest we stick with one top-level describe per module to work
around this issue.