service session meta data: sketching the task.
Helloworld should be able to ask thentos for information about the
session is it serving, such as the name of the user or session timeout.
(It should also be able to update that metaata, but that's a different
story.)

This is the sketch of a pull request.  To complete it, here are some
hints as to how to proceed.

1. re-align everything with the new types without changing functionality.
2. implement trans_getServiceSessionMetaData, and expose via rest api.
3. extend helloworld service to connect to thentos via rest api and call
   the new transaction.

Come to think of it, these are probably three separate pull requests.