Dead end.
This is just to see if we can add the error type as a second class
param rather than in a type family.

Effects: we suddenly need lots of proxies.  This is at least ugly,
but probably worse, since this sinks the quantifiers in the Lens'
type into the function type and thus requires Rank2Types, with
bad consequences when trying to get Thentos.Transaction.Transactions
to type-check again.

I think we shouldn't really do this.  I'll commit it for refernce,
then revert.

NOTE: The changes in src/Thentos/Transaction/Transactions.hs are
even more preliminary than the ones in Types and Core.