Explicitly bind to the address we connect to.
Before, the proxy and destination servers would bind to '*4'
(compare Data.Streaming.Network(HostPreference)),
which could succeed despite a separately running server binding
more closely(?) to 127.0.0.1, so that later on requests to
http://localhost would hit the other server...

This doesn't feel like the right fix, but I'm not sure what is.

At any rate, with this, I get an actual error message (though
no test failure) when I run

$ nc -l localhost 8001

on the side (OS X).