Implement password_reset endpoint as per the A3 spec
This endpoint is invoked after the A3 backend has been triggered (via the
create_password_reset endpoint) to send a password reset mail.

We forward the request (but without the new password) to the backend to
find out whether the reset path is valid and to which user it refers.

If it's valid, we change the password in Thentos.

As per the spec, the user is afterwards logged in and not-yet-activated
users become activated.