Remove serviceGroups field from Service
I'm not entirely sure what this was for, but we're not currently using it
and the way it exists now will have to be reconsidered anyway if we want
to use it with a relational db. I think removing it is better than leaving the
field in and just drop the data when saving the service to the db.