Remove unused thSessServiceSessions field from ThentosSession type
Up until now, this field was always empty and when we got a thentos session
from the db (meaning that the actions that relied on this data were broken).
This field doesn't really make sense in a relational context, so I removed it
and fixed the affected actions (just serviceNamesFromThentosSession) by
implementing the necessary transactions.