safer ToField instance for Timeout
The old ToField instance would convert 3 seconds to '3.0',
with this it's converted to '3.0 seconds'. It's less easy
but not impossible to accidentally misuse this value as a non-interval.

Not quite sure about this. It was part of an attempt to fix
errors when providing an interval to 'now() + ?', which just
complains about the decimal point in the floating point literal.