Fix call to grantAccessRights'P.
grantAccessRights'P does not look up roles of an agent, so it
should be considered an internal function in general.  (admittedly,
there is a lot of potential for cleaning up the code involved.  i
apologize for that.)

in order to make this error a little harder (and for catching all
the occurances of it that were already in the code), i restricted
the involved types a little.