Add a config field that describes where a3 thinks its backend api lives.
This is a URL prefix, i.e. a valid URL that will be prepended to e.g. user
paths. So if the prefix is http://example.com/ , we expect the full user path
to be http://example.com/principals/users/<id> .
This is important because a3 doesn't handle deviations from the expected
user paths very well, so we have to make sure that e.g. the value of
X-User-Path matches the expected format exactly (i.e. has the same protocol,
doesn't have a redundant port number in the URL etc.).